New event (#52)
Closes #48 #21

This great simplifies the events, by putting common code in the base class and removing the duplicates in the derived classes. The event is conceptually a non-owning view into the Fortran record, which you should not modify in-place. The filtering logic with complicated caching was removed, which attempted to give the illusion of being able to apply filter in place while keeping the original record intact.

Instead, the event now uses the familiar numpy syntax for applying masks `event[mask]`, which in general returns a copy (same as with the previous filtering logic). This is now explicit, instead of implicit, which is better design. Applying masks on the event returns a simpler EventData object, which intentionally does not have all the methods of the original event and which in general is a copy. The original event is now read-only, to ensure that the record is not accidentally modified. The changes greatly simplify the implementation.

The `npart` attribute was dropped from the event, instead the more pythonic `len(event)` may be used. 

An event can now be copied with `event.copy()`, which produces a `EventData` object that can be pickled, unlike the original MCEvent. This is very useful, because it allows one to store impy events with pickle or other serialization libraries, and also very important: allows one to return impy events from processes started by `multiprocessing`. I use this extensively in the unit tests to work around the initialization limitation.