Add more tests to test_to_hepmc3.py and test_hepmc_io.py (#74)
This patch tries to find the root cause of the failures in `test_hepmc_io.py`. It
expands `test_to_hepmc3.py`, because at least some failures originate
from invalid parent records in the HEPEVT, and those are easier to
expose in `test_to_hepmc3.py`. It is not a fix, though.