Pythia8 (#77)
**Warning**

When compiling this on your machine, you may face multiple issues if you
have very recent libraries.
- setuptools > 60 does not seem to work. Downgrade with `pip install
setuptools<60`
- If you have OSX: Apple introduced a bug in the latest Command Line
Tools 14 which produces a linker error when compiling C++ code with gcc.
Many projects are affected. The only workaround is to downgrade to 13.4,
use this link
https://download.developer.apple.com/Developer_Tools/Command_Line_Tools_for_Xcode_13.4/Command_Line_Tools_for_Xcode_13.4.dmg
and turn off automatic updates in the System Settings, because otherwise
your Mac will upgrade to 14 again.

Closes #76 

This adds Pythia8 and is a major cleanup of the functionality in
existing models and common.py. While adding Pythia, I found many other
things that were worthwhile to fix along the way, so this became a large
patch. I fixed the IO issues for all models and removed a lot of
duplicated code through refactoring. Pythia8 is the first model which
can be safely restarted, which means that you can generate as many
instances of the Pythia8 class in the same interpreter as you like.

Custom bindings are generated with pybind11. I don't use the Python
bindings that Pythia8 provides, because those are autogenerated and I
don't fully trust that autogenerated bindings make the correct decisions
about object ownership. Another reason is that I am building Pythia8
through our CMake file, which ignores the original build system of
Pythia8, but the original build system was generating these bindings on
the fly. Writing the small set of bindings that we actually use was
easier for me than figuring out how to make all that work. If we start
to use large parts of Pythia8 API and find out that having customized
bindings does not give us significant gains, then this decision should
be revisited.

pybind11 is included as a submodule. Pythia8 comes bundled with
pybind11, but the copy in Pythia8 lacks the CMake files that we need.

Other changes:
- `mold` linker is used during build if available. `mold` is supposed to
be much faster than the standard linker.
- `impy.definitions` was removed. The settings in `impy.definitions` are
class attributes now. The advantage of this is that one can query
information about the model class, without creating an instance, e.g.
```py
from impy.models import Sibyll21, Sibyll23d

Models = [Sibyll21, Sibyll23d]

for Model in Models:
     print(Model.name)  # this used to fail, because Model is a class not an instance
```
- `EventData.__getitem__` now returns a reduced event **which still
contains a parent history**. A new algorithm in
`impy.util.select_parents` shifts the parent indices after selection.
This algorithm is optionally accelerated with numba if numba is
available. This was a pain to develop and debug, but the solution I have
now is short and generic.
- `EventData.to_hepmc3` now returns a valid GenEvent for all models and
all models can be written with pyhepmc thanks to that. This comes at the
cost of reducing the Phojet and UrQMD records to their final state
during the conversion, and in case of Pythia-8 I have to throw out the
parton shower (but decay history is kept, which is what 99 % of people
really need). Warnings are emitted for those models which cannot convert
the full event to HepMC3. If the underlying issues can be fixed, we can
remove these workarounds. I am not sure that this is worthwhile for
Phojet, but it may be useful to have a proper history for UrQMD.
- `common.Settings` was removed. We don't use this and we don't need an
abstract base class for settings.
- `dpmjetII.py` was removed. It contains only dead untested code.
- `Model.lib` renamed to `Model._lib`
- `Model.attach_log` renamed to `Model._attach_log`
- `Model.generate_event` renamed to `Model._generate_event`
- `MCRun._define_default_fs_particles()` renamed to
`MCRun._set_final_state_particles()`
- `Model.init_generator` converted into `Model.__init__`
- `MCRun.set_stable` now does common logging and calls
`Model._set_stable` to do the actual work
- Common code used during Model initialization was moved to
`MCRun.__init__`
- `Model.sigma_inel` generally accepts `evt_kin` as an argument now.
`Model.sigma_inel_air` does not accept any arguments. The latter should
be removed in the long-term, since we have the composite target for
this.
- For nuclei, event_kinematics returns the appropriate p1pdg and p2pdg.
Code duplication was reduced.

Co-authored-by: afedynitch <afedynitch@gmail.com>
Co-authored-by: Anatoli Fedynitch <afedynitch@users.noreply.github.com>