add CLI that mimics CRMC (#89)
First step towards a CRMC-like cli. This is a basic version, but it
already has reasonable compatibility with crmc. It can generate HepMC
files, Root files, and SVG images of events (useful for debugging).

Closes #86 

Related issues to be addressed in future PRs:
- #92 
- #93
- #94

Other major changes
- A new dataclass CrossSectionData was added to common.py. It allows us
to store multiple cross-sections at once (total, elastic, inelastic,
several diffractive).
- Model.sigma_inel was replaced by Model.cross_section, which calls
Model._cross_section to do the actual work
- Model._cross_section is completely filled for Pythia6 and Pythia8, and
nearly for EPOS-LHC and SIBYLL-2.3
- The code to generate events for composite target was refactored to
make it simpler and more clear. I think the implementation (that I have
not changed) is not correct, because ecm is not changed when the nuclear
partner changes. The composite target probably only works correctly if
the model frame is the fixed-target frame.
- EventData has two new fields: impact_parameter and n_wounded
- The properties n_wounded_A, n_wounded_B, n_spectators_A, and
n_spectators_B were removed from the models that had them; n_wounded_X
was replaced with EventData.n_wounded which returns a tuple;
n_spectators_X was completely removed because it can be computed from
number of nucleons - number of wounded. We can add properties that
compute these numbers if needed. Storing redundant numbers is bad.
- The set_unstable bug in EPOS was fixed (see #86)
- particletools was removed as a dependency
- The special case for the pi0 stability in the config was removed. We
should have only one mechanism to declare which particles are stable and
that is the stable_list.
- Lots of new tests