Remove config (#90)
**This is a large refactoring and cleanup and thus should be merged
before any other new PRs are started**.

Closes #13 
Closes #31
Closes #85 

impy.config is removed. We need a config file for the CLI that mimcis
the crmc config file, but that's another separate issue.

I fixed bugs along the way and decluttered the interfaces, reduced code
duplication and more.
- OutputGrabber and _attach_log methods and friends have been removed.
We need a different solution to catch output from the Fortran codes. I
have some ideas, which will be explored in future PRs. For now, all
models print to stdout.
- Models now use a unified system to check whether input particles are
valid, based on attributes `_projectiles` and `_targets`. Models can
override the defaults set in MCRun.
- Models also use a unified system to check whether the requested
kinematics is valid. Most models have a minimum sqrt(s) of 10 GeV. We
need to catch this in Python, because FORTRAN typically calls STOP when
the energy is too low, which aborts the Python interpreter or kills the
Jupyter kernel.
- CompositeTarget had a separate random number generator and a separate
seed, that was bad, because we need a single source of randomness if we
want runs to be reproducible. It still uses a seperate random number
generator, but at least the seed is now sync'ed with that of the model.
Generally speaking, all random sampling must happen in MCRun and should
go through RMMARD. This is currently not the case when you use a
CompositeTarget.
- The output frame is now set via the `EventFrame` enum class, instead
of a string.
- The output frame is defined by the input frame and stored in
`EventKinematics`. Users can override the frame to get the output in
another frame as the input. If the model supports output in different
frames, we use that functionality, otherwise boost ourselves in Python.
The model also has a frame, but this is the frame in which the output is
actually generated, not the frame that the user requested. Users can
change the frame in `EventKinematics`, but in case of the generator it
must be read-only. The output frame was previously a mutable global
variable, shared between all models, which was a source of potential
bugs.
- `EventKinematics` stores the particles as `PDGID` from `particle`
instead of a plain int, because `PDGID` has useful properties, like .Z
and .A and .is_nucleus that we duplicated before. This allows us to drop
some duplicated code and related (not well-designed) API.
- `MCRun.event_kinematics` was shortened to `MCRun.kinematics`
- `MCRun._generate_event` was shortened to `MCRun._generate`
- `MCRun.cross_section` now also accepts a `CompositeTarget` and
computes the correct cross-section.
- EPOS initialization was simplified, some variables were initializated
twice. Frame is now set in `_set_kinematics`, because a new
`EventKinematics` may require changing the output frame.
- I removed the option to give names to components in the
`CompositeTarget` names. This is not adding value and just makes the
code more complex. I also removed the API to add components, which was
just making the design more complicated. The user can always generate
the list of components and fractions with standard Python collections
then pass the list to the constructor of `CompositeTarget`. We should
not provide an API for trivialities, especially when it makes our design
more complicated (it was not clear when to normalize the input fractions
when incrementally adding components is allowed).
- `test_generators.py` now checks generation with composite targets and
all supported output frames. The check is much more elaborate now by
including functionality previously in `fixme_all_cms.py`,
`fixme_all_lab.py`, `test_composite_target.py`, which were all removed.
A plot is automatically generated for visual inspection locally and
uploaded to GitHub when the test runs in CI.
- `test_setstable.py` (formerly `fixme_setunstable.py`) was added and
improved, now it tests all models and exposes those which do not support
setstable.
- `fixme_eposlhc.py` was removed, because it was obsolete

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Anton Prosekin <ayprosekin@gmail.com>
Co-authored-by: afedynitch <afedynitch@gmail.com>