Replace RMMARD with NumPy RNG (#115)
Closes #102 
Closes #104 

Any Numpy PRNG can be used (not only PCG64), currently it is the default
one from Numpy. There should be no need to use any other. Seeds are
handled completely in Python by the base class MCRun, Fortran codes
should not set seeds. The Fortran functions which do any kind of RPNG
initialization have been replaced with noop stubs.

I performed some crude benchmarks to see whether this change impacts the
speed of the generators. SIBYLL-2.1 is the fastest, it got maybe 5 %
slower. EPOS is one of the slowest models, it got 13 % faster.

Benchmark: Runtime to generate 10000 pp events at 1 TeV and write as
hepmc3.

|                       | EPOS-LHC  | Sibyll-2.1 |
|--------------|-------|----------|
| main              | 51 s     | 3.1 s         |
| this branch   | 45 s     | 3.3 s        |

test_rng_state fails for EPOS and UrQMD. EPOS used to pass. We can fix
this in a separate PR. I did some investigation but could find out why
this is happening.

gasdev and spgasdev (used by the Sibylls) are now implemented in terms
of the numpy random normal generator. This allows us to get rid of the
extra state which the old implementations held. If we want reproducible
randomness, we need to get rid of random functions which cache some
results.

**Other changes**
- In addition to extra.cfg, we now also have model.cfg, which contains
the models that are normally build. When changing Fortran code, it can
be useful to comment out models which are not affected during
development here. CMake will then not build those models.
- I fixed a bug in F2PY.cmake, the compile definitions were not passed
correctly to numpy.f2py.
- I removed the obsolete Makefiles.
- Renamed src/fortran/EPOS-LHC to src/fortran/epos to be more consistent
with other models.
- Fixes a bug in the CLI: seed was never passed to the model!
- Added a null writer for the CLI, which is useful for benchmarking.
- I deactivated test_generators on the CI Windows runner. It is enough
to test all models twice on MacOS and Linux, and the Windows tests
generally take much longer than the others. Skipping this should make
the runtimes more even.
- I skip test_generators with Model=UrQMD and frame=cms2ft and p2=p,
since this test is generally flaky.

Co-authored-by: afedynitch <afedynitch@gmail.com>