Epos random dummy (#121)
Minor change.

After talking to Tanguy, I agree that it is better to reduce ifdefs in
the original fortran codes. So instead of using ifdef CHROMO in
epos-random.f, I made a dummy epos-random-dummy.f with the blank
initialization routines and implemented the functions rangen and drangen
that produce random numbers in rangen.fpp.