chromo logo (#123)
Closes #2 

Here is a logo in metallic font. I made this myself with inkscape. The
metal effect uses three colors blue, green, and red, which represent the
three color charges of QCD. The font I used is Nova Mono, which is
released under a free software licence that allows this kind of use.