Simpler and more efficient Pythia8 event handling (#124)
The Pythia8 event is now a view into the original Pythia8 event
structure. Copies are avoided and the new design allows us to use
several independent Pythia8 instances in parallel, since there are no
global shared data structures. In summary, the event access is faster
and the code is more straight-forward, too.

Since direct views require access to private members, I once again use
cpp-member-access, which uses a loophole in the C++ standard to do just
that. It is safe and works on all platforms, I already use this in
iminuit and pyhepmc.

This PR builds on #120 which should be merged first.