Hot-fix for broken second parent in Pythia8 parents (#140)
This is a hot-fix for my recent change to the Pythia8 event view.
Undetected by our tests, the second parent index was pointing to the
wrong memory position, so that the second parent index was garbage. I
added a new test to which should catch this in the future.

This is another hot-fix for a bug, so I will merge this myself.