Add CRMC performance comparison (#138)
I wanted to check how the results that @jncots got a while ago on the
performance comparison between CRMC and chromo look when we use the
current dev version of Chromo. I cannot reproduce the large speed-ups of
up to 35x. In my comparison, chromo is faster for QGSJet-II-04 and
SIBYLL-2.3d and slower for EPOS-LHC. For full transparency, I added the
notebook with my code to the folder `bench`.


![image](https://user-images.githubusercontent.com/2631586/221798975-4f58658c-1eaa-47de-8365-690aa135685f.png)

The results look reasonable, because I don't expect chromo to be faster
by an order of magnitude. After all, both run the same Fortran codes and
that should dominate the overall CPU load. I expect some differences
from the random number generator and compiler options, and
postprocessing. But these should be all small effects. In the original
comparison, was CRMC perhaps compiled in DEBUG mode but Chromo in
release mode? This could explain the large differences in speed.

Some general notes:
- The comparison depends on the machine where the benchmark is run. We
are testing a mix of CPU and disk load. My computer has a harddrive,
results with a SSD should be different. I tried to take that out in my
benchmark by writing to /dev/null, which is a virtual file that just
discards arriving blocks.
- For a meaningful comparison, we need to make enough events per
generator so that the startup time becomes negligible.

---------

Co-authored-by: Anton Prosekin <ayprosekin@gmail.com>