Building wheels on CI (#82)
The main goal of this PR is to fix the problems which appear on CI when
binary wheels are built using `cibuildwheel`. The main artifact is
`Release` workflow in `release.yml` file which describes the process of
building wheels and uploading them on PyPI.

The following problems were solve (in the order of their solution):

* Building wheels for Windows. It is still not solved how to build
`Pythia8` and `urqmd34` (see discussions below). So currently these
models are excluded from the build. The main trick to solve the problem
was adding the following lines in `F2Py.cmake` file:

```
if (WIN32)
    target_link_libraries(${target_name} PUBLIC "-static")
 endif()
```
Also there was a problem with setting correct generator in CMake and
finding `gfortran`.

* The main problem with Linux was that in Pythia8 there are 2 files:
_pythia8 and libpythia8. _pythia8 depends on libpythia8. However, during
build time `_pythia8` obtained rpath, which doesn't allow to find
libpythia8, if both files in the same directory. It was fixed by adding
`LINK_FLAGS "-Wl,-rpath...` for `_pythia8` target in CMakeLists.txt.

* With MacOS there were 2 main problems: set correct version of
`gfortran` and `c++` and linking with Python library.
- The solution of the problem with compilers was copied from
`https://github.com/numpy/numpy/tree/main/tools/wheels`.
`gfortran_utils.sh` script was copied from numpy to install the correct
version of gfortran.
       
     - The problem with Python library was finally solved in PR #126

* The last problem solved which is worth to mention is the finding a
wrong python version via find_package(Python). The key issue was that
`cibuildwheel` uses venv but do not activate it. Because of that
find_package(Python) could not find python used in virtual environment.
It is solved by setting environment variable VIRTUAL_ENV in setup.py

---------

Co-authored-by: Hans Dembinski <hans.dembinski@gmail.com>
Co-authored-by: Hans Dembinski <HDembinski@users.noreply.github.com>
Co-authored-by: afedynitch <afedynitch@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Anatoli Fedynitch <afedynitch@users.noreply.github.com>
Co-authored-by: Henry Schreiner <HenrySchreinerIII@gmail.com>