Fix workflow file 2 (#149)
This is a second trial to launch PyPI. There is a problem with
conditions. It seems that `github.event.base_ref` is null. At least this
is what it shows when the workflow was launched on a fork. On my test
repository `github.event.base_ref` has correct value. I tested
conditions on a fork of chromo and made a release. It worked. I hope
this time it should work.