Gamma+p example (#146)
The Leptohadronic modelling workshop made me curious what needs to be
done so that Pythia8.3 supports gamma+p interactions like SOPHIA. Turns
out, Pythia8.3 as implemented in Chromo already supports that. I added a
simple example which compares SOPHIA with Pythia-8.3. This can be used
to advertise Chromo to the workshop participants.


https://github.com/impy-project/chromo/blob/1546cf9c8cfdc9cc189b0dda7bf35c1bf92d1fe1/examples/gamma_p.ipynb

Comments and feedback are welcome.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>