add refill_decay_stack (#161)
When Pythia8 is used for decay of particles, it is required to put the
particle that should be decayed on a `event` stack . Currently it is
only possible to do like this:

```python
for i in range(len(stack)):
    pid = stack.pid[i]
    en = stack.energy[i]
    m = self._pythia.particleData.findParticle(pid).m0
    pz = sqrt((en - m)*(en + m))
    self._pythia.event.append(pid, 91, 0, 0, 0, 0, pz, en, m)
```

So, to put a lot of particles from vector stack, it is needed to call
`append` for every particle. This is time consuming because of Python
`for` loop and the call to `append` itself. The time of filling loop is
the same as the time required for decay in Pythia, which from
calculation point of view are not comparable tasks at all.

The added function `self._pythia.refill_decay_stack(stack.pid,
stack.energy)` accepts 2 numpy arrays with pid and energy, and calls the
above loop in C++. These significanly (~40 times for my test cases)
reduces the time for filling the `event` array.