Test of final state (#160)
This PR fixes the issues with DPMJET and PHOJET models, which didn't
follow the decay settings controlled by chromo, and adds a test which
checks that all particles which we expect to occur in the final state
actually occur for all models.

Two exceptions are built into the test:
- SIBYLL-2.1 does not generate Omega- (and cc)
- UrQMD doesn't generate only Omega- (however this might be related to
limited stats)
- The QGSJet family does not generate Omega-, Xi0, Xi-, Sigma+, Sigma-
(and cc)

The PR fixes the following problems to make the test pass:
- [x] UrQMD does not generate KS particles. Bug was fixes in our fortran
middle layer.
- [x] DPMJetIII and PHOJET were not following the decay settings
requested. Bugs have been identified and fixes to DPMJET backported +
submodules updated.
- [x] Vitange DPMJET/PHOJET had the problem as well, and it was fixed in
`chromo`. The minor version of both codes have been upped by one to
reflect the minor change.
- [x] All tests should pass with less XFAILs for decays in PHOJET/DPMEJT
models

This PR closes #156.

---------

Co-authored-by: afedynitch <afedynitch@gmail.com>