Maintenance release (#189)
The PR is a maintenance release with the following changes:

1. From version 1.26.2 we can build the f2py interface files natively
with the public `numpy` release with no hacking. This has been reflected
in the build chain.
2. A quantity "called `fw` that was need for producing histograms of the
invariant cross section `E*dsigma/dp3` has been brought back to the
event data. It requires the center of mass momentum `pcm` to be
accessible from the `EventKinmatics` class (same as for `ecm`).
3. Repair event history and pre-/appending beam particle info can now be
optionally disabled by setting `generator._restore_beam_and_history =
False` to save some CPU time when it is not needed. By default it
remains on.
4. When subclassing `MCRun`, for example when developing custom
interaction models, the model library may not be called
`chromo.models.library_name`. To get this flexibility, the `MCRun` base
class is searching under the previous path `chromo.models` for library
name and if it fails tries just `library_name`.
5. `__hash__` for EventKinematics such that instances can be dictionary
keys.
6. Fixed bugs in QGSJET cross section function when it was called with
nuclear or meson primary. Fixed major bug in QGSJET-II cross section
readout. `qgsect` supports only 3 types of primaries and not all the
known particles.
7. Cross section readout fixed for SIBYLL with meson primaries (where
p.A is None)
8. Implemented missing cross section readout for SIBYLL 2.1
9. Some cosmetics and some imports sorted